Fix unbuildable desktop sample due to missing kflite
