+++ Sample/src/desktopMain/kotlin/org/company/app/Util.desktop.kt
@@ -1,0 +1,5 @@
+package org.company.app
+
+import kotlinx.coroutines.CoroutineScope
+
+actual fun getTFliteRunner(): (ByteArray.(CoroutineScope) -> Unit)? = null
